% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_results.R
\name{format_results}
\alias{format_results}
\title{Format SuSiEx Output Files}
\arguments{
\item{path}{A character string specifying the directory containing the SuSiEx output files. The user must check there are no duplicate files of fine mapped regions.}

\item{ancestries}{A character vector specifying the ancestries used in the SuSiEx analysis.}
}
\value{
A list with the following components:
  - `summary`: A data frame containing the combined data from all `.summary` files.
  - `cs`: A list of data frames for `.cs` files, each annotated with position information.
  - `snp`: A list of data frames for `.snp` files, each annotated with position information.
}
\description{
Processes the SuSiEx output files by extracting position information from `.summary` files and
applying it to `.cs` and `.snp` files. Files that cannot be processed are excluded.
}
\examples{
\dontrun{
format_results("path/to/susiex/results", ancestries = c("EUR", "AFR"))
}

}
