% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAncestryCausal.R
\name{plotAncestryCausal}
\alias{plotAncestryCausal}
\title{Plot Ancestry-Specific Causal Inference Probabilities}
\arguments{
\item{summary_results}{A data frame containing the summary results from a
SuSiEx analysis, the object returned from format_results()$summary.}

\item{ancestries}{A character vector containing the names of ancestries
present in the summary results, used to separate the ancestry columns and
assign them to the correct columns in the output.}
}
\value{
A ggplot2 object containing a scatter plot of the post-hoc
probabilities for SNPs, showing causal inference probabilities across
different ancestries.
}
\description{
This function generates a plot visualising the post-hoc probability of causal
SNPs across multiple ancestries. It processes the .summary results from a
SuSiEx output, organises the ancestry data, and produces a scatter plot with
the probability of being causal (POST-HOC_PROB_POP_CS) for each SNP across
different ancestries.
}
\examples{
\dontrun{
plotAncestryCausal(results$summary, c("EUR", "AFR", "SAS"))
}

}
