% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPurityPIP.R
\name{plotPurityPIP}
\alias{plotPurityPIP}
\title{Plot Relationship Between Credible Set Characteristics and PIP}
\arguments{
\item{summary_results}{A data frame containing summary results, which should include the columns `CS_PURITY`, `MAX_PIP`, `CS_LENGTH`, and `CHR`.}
}
\value{
A ggplot2 object that visualizes the relationship between `CS_PURITY`, `MAX_PIP`, and `CS_LENGTH` across chromosomes.
}
\description{
This function generates a scatter plot showing the relationship between three key characteristics of credible sets:
- `CS_LENGTH`: The number of SNPs in the credible set.
- `CS_PURITY`: The purity of the credible set.
- `MAX_PIP`: The maximum posterior inclusion probability (PIP) in the credible set.
}
\details{
The plot visualizes the data as follows:
- The x-axis represents `CS_PURITY` (the purity of the credible set).
- The y-axis represents `MAX_PIP` (the maximum posterior inclusion probability).
- The size of each point corresponds to the number of SNPs in the credible set (`CS_LENGTH`).
- Each point is colored by chromosome (`CHR`).
}
\examples{
\dontrun{
plotPurityPIP(results$summary)
}

}
