% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_file.R
\name{process_file}
\alias{process_file}
\title{Process Individual SuSiEx Output Files}
\usage{
process_file(file, file_ext, chr_info = NULL, ancestries)
}
\arguments{
\item{file}{A character string specifying the path to the file to be processed.}

\item{file_ext}{A character string specifying the extension of the file ("summary", "cs", or "snp").}

\item{chr_info}{A named list containing chromosome (CHR), start position (BP_START), and end position (BP_END)
extracted from the `.summary` file, required for `.cs` and `.snp` files. Default is `NULL`.}
}
\value{
A data frame containing the processed results, or `NULL` if the file could not be processed.
}
\description{
Reads and processes individual SuSiEx output files, extracting chromosome and position data from the `.summary` file
and using that information to process `.cs` and `.snp` files.
}
\keyword{internal}
