% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_file.R
\name{process_file}
\alias{process_file}
\title{Process Individual SuSiEx Output Files}
\usage{
process_file(file, file_ext)
}
\arguments{
\item{file}{A character string specifying the path to the file to be processed.}

\item{file_ext}{A character string specifying the extension of the file (either "summary", "cs", or "snp").}
}
\value{
A data frame containing the processed results, or `NULL` if the file could not be processed (e.g., if it has no results or is invalid).
}
\description{
This function reads and processes individual files from SuSiEx output, such as `.summary`, `.cs`, and `.snp` files. It extracts relevant information, including chromosome and position data, and handles file reading and error management.
}
\details{
- For `.summary` files, it reads the data skipping the header line.
- For `.cs` and `.snp` files, it reads them directly as data frames.
- If the file has no rows or contains only two columns (indicating that it is empty or invalid), the function returns `NULL` and skips processing that file.
- It extracts chromosome (CHR), start position (BP_START), and end position (BP_END) from the file name using regular expressions.

This function is called internally by other functions in the package to process multiple files at once.
}
\keyword{internal}
